moving error classes/utilities to common module
