+++ pydough/errors/__init__.py
@@ -1,0 +1,17 @@
+"""
+Module for error handling in PyDough.
+"""
+
+__all__ = [
+    "PyDoughMetadataException",
+    "PyDoughQDAGException",
+    "PyDoughTypeException",
+    "PyDoughUnqualifiedException",
+]
+
+from .error_types import (
+    PyDoughMetadataException,
+    PyDoughQDAGException,
+    PyDoughTypeException,
+    PyDoughUnqualifiedException,
+)
