+++ pydough/metadata/__init__.py
@@ -8,7 +8,6 @@ __all__ = [
     "GeneralJoinMetadata",
     "GraphMetadata",
     "PropertyMetadata",
-    "PyDoughMetadataException",
     "SimpleJoinMetadata",
     "SimpleTableMetadata",
     "SubcollectionRelationshipMetadata",
@@ -17,7 +16,6 @@ __all__ = [
 ]
 
 from .collections import CollectionMetadata, SimpleTableMetadata
-from .errors import PyDoughMetadataException
 from .graphs import GraphMetadata
 from .parse import parse_json_metadata_from_file
 from .properties import (
