+++ pydough/metadata/graphs/graph_metadata.py
@@ -2,8 +2,9 @@
 Definition of PyDough metadata for a graph.
 """
 
+from pydough.errors import PyDoughMetadataException
+from pydough.errors.error_utils import HasType, is_valid_name
 from pydough.metadata.abstract_metadata import AbstractMetadata
-from pydough.metadata.errors import HasType, PyDoughMetadataException, is_valid_name
 
 
 class GraphMetadata(AbstractMetadata):
