+++ pydough/metadata/parse.py
@@ -6,18 +6,19 @@ __all__ = ["parse_json_metadata_from_file"]
 
 import json
 
-from .collections import CollectionMetadata, SimpleTableMetadata
-from .errors import (
+from pydough.errors import PyDoughMetadataException
+from pydough.errors.error_utils import (
     HasPropertyWith,
     HasType,
     NoExtraKeys,
-    PyDoughMetadataException,
     extract_array,
     extract_bool,
     extract_string,
     is_json_object,
     is_string,
 )
+
+from .collections import CollectionMetadata, SimpleTableMetadata
 from .graphs import GraphMetadata
 from .properties import (
     CartesianProductMetadata,
