+++ pydough/metadata/properties/scalar_attribute_metadata.py
@@ -7,8 +7,8 @@ __all__ = ["ScalarAttributeMetadata"]
 
 from abc import abstractmethod
 
+from pydough.errors.error_utils import HasType
 from pydough.metadata.collections import CollectionMetadata
-from pydough.metadata.errors import HasType
 from pydough.types import PyDoughType
 
 from .property_metadata import PropertyMetadata
