+++ pydough/qdag/collections/collection_qdag.py
@@ -13,8 +13,8 @@ from typing import Union
 
 import numpy as np
 
+from pydough.errors import PyDoughQDAGException
 from pydough.qdag.abstract_pydough_qdag import PyDoughQDAG
-from pydough.qdag.errors import PyDoughQDAGException
 from pydough.qdag.expressions.collation_expression import CollationExpression
 from pydough.qdag.expressions.expression_qdag import PyDoughExpressionQDAG
