+++ pydough/qdag/collections/partition_by.py
@@ -9,8 +9,8 @@ __all__ = ["PartitionBy"]
 
 from functools import cache
 
+from pydough.errors import PyDoughQDAGException
 from pydough.qdag.abstract_pydough_qdag import PyDoughQDAG
-from pydough.qdag.errors import PyDoughQDAGException
 from pydough.qdag.expressions import (
     BackReferenceExpression,
     ChildReferenceExpression,
