+++ pydough/qdag/collections/where.py
@@ -8,7 +8,7 @@ __all__ = ["Where"]
 
 from functools import cache
 
-from pydough.qdag.errors import PyDoughQDAGException
+from pydough.errors import PyDoughQDAGException
 from pydough.qdag.expressions import PyDoughExpressionQDAG
 from pydough.qdag.has_hasnot_rewrite import has_hasnot_rewrite
