+++ pydough/qdag/expressions/back_reference_expression.py
@@ -4,8 +4,8 @@ context.
 """
 
 __all__ = ["BackReferenceExpression"]
+from pydough.errors import PyDoughQDAGException
 from pydough.qdag.collections.collection_qdag import PyDoughCollectionQDAG
-from pydough.qdag.errors import PyDoughQDAGException
 from pydough.types import PyDoughType
 
 from .expression_qdag import PyDoughExpressionQDAG
