+++ pydough/types/array_type.py
@@ -6,7 +6,8 @@ __all__ = ["ArrayType"]
 
 import re
 
-from .errors import PyDoughTypeException
+from pydough.errors import PyDoughTypeException
+
 from .pydough_type import PyDoughType
